backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==21.2.0
flask-socketio==5.3.6
eventlet==0.33.3
pytz==2023.3

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7

# Notifications & Scheduling
APScheduler==3.10.4
sendgrid==6.11.0
